Updated SQLSelectQuery code for non-static Property IDs
Updates the assembly to look for GlobalAction properties called
"Connection String", "SQL Statement", and "Return Value" instead of
looking for properties 1, 2, and 3 respectively. This will allow the
creator of a GlobalAction workflow to not have to recreate their
workflow if they forget to create the properties first.

Included: Rebuild of DLL file.